Many changes format_pqtl_ukbppp - the main change was that the rsid file has column names tagged with the rsid_file ending to prevent "column_name".x and "column_name".y from being formed which affect the harmonisation function
run_mr - main changes are I started to delete code which is redundant. There are still more pieces of code which can be deleted within the function. I decided that the format_data() for the outcome data frame needs to be within the function because that makes the most sense. However, the column names for that data frame should be altered before the function is run